requirements.txt
# Core dependencies for RooCode local agent system
pyyaml==6.0.1
click==8.1.7
requests==2.31.0
jinja2==3.1.2
jsonschema==4.19.2
python-dotenv==1.0.0

# Testing and validation
pytest==7.4.3
pytest-cov==4.1.0
yamllint==1.32.0
ruff==0.1.6

# Optional LLM integration
transformers==4.35.2
torch==2.1.1
llama-cpp-python==0.2.11

# Development tools
setuptools==68.2.2
wheel==0.41.3
pip==23.3.1
